refactor: move jwt_config package to remove import cycle
