+++ internal/util/jwt/jwt_config.go
@@ -1,23 +1,22 @@
-package config
+package jwt
 
 import (
 	"crypto/rsa"
 	"sync"
 
 	"github.com/Team-OurPlayground/our-playground-auth/internal/util/customerror"
-	"github.com/Team-OurPlayground/our-playground-auth/internal/util/jwt"
 )
 
 var (
-	JWTOnce    sync.Once
+	once       sync.Once
 	privateKey *rsa.PrivateKey
 	publicKey  *rsa.PublicKey
 )
 
 func InitJWTKeys() {
-	JWTOnce.Do(func() {
+	once.Do(func() {
 		var err error
-		privateKey, publicKey, err = jwt.InitRSAKey()
+		privateKey, publicKey, err = InitRSAKey()
 		if err != nil {
 			panic(customerror.Wrap(err, customerror.ErrInternalServer, "config InitJWTKeys error"))
 		}
